changed almost everything
this is able to call Jitsi, and recognizes several failure cases (e.g. busy), and the general call failure case. this is also able to establish a session with Jitsi, though it does not actually send SDP in the INVITE, so Jitsi seems to fill it in with its own choice. the ACK to the INVITE 200 does not confirm the choice of media yet, since we aren’t even sending any (so this is not actually following the legal process for setting up a session, though Jitsi seems to have our back here).

added a new interface to tell users of SipClient about actual SIP events. the text interface is really just for testing.

there are a huge number of cases still unhandled even in the parts that are basically working. for major things, handling an incoming call should probably be the top priority, followed by BYE, followed by SDP.

this should probably start to be decomposed further now. e.g. a Call abstraction might be a good idea.