can mostly handle empty calls
Can accept and decline calls most of the time (may still encounter that race condition occasionally...? see SipResponder), and the other side seems to get the message. This seems to apply for both Jitsi and ourself.

This turned out to be  complicated to achieve even when it seemed very close; telling Jitsi that we accepted/declined was easy, but then sending sensible messages to ourself turned out to be hard. One race condition was my fault, due to poor locking logic. But a related issue is still JAIN's fault, and this person makes several good points near the end of "Transactions": https://technfun.wordpress.com/tag/jain-sip-api/
On that note, I'm still not sending or handling CANCEL

I really need to factor out call state into RTTCall. It's getting ugly, and i just made it worse.