sending some real text
Progress on actual text detection. When characters are purely added, this is mostly detected OK. Some extra nonprinting character seems to be appended, maybe when spaces are involved? Oh, that has transformed into a backspace, it seems, so it deletes a char on the other end. Any removal of characters is treated as exactly one backspace for now, and sometimes more than one accidentally seems to be sent?