unrelated to this actual branch
these changes make ACK work better, amazing that i didn't have this problem until now.

but this branch's purpose, the dummy audio, is no good. asterisk doesn't buy it. it allows the call to go through, but disconnects after 30sec when no audio is detected.

i suppose i can still leave it in, since it works, but just disable it.