direct media
Bizarrely, we were not sending media directly to each other before, as specified in the SDP. We were using the Contact header's address, which I guess is only meant for SIP messages. We should use the IP in the SDP, so now we are. Surprising that it took this long for this to be a problem. Normally the Contact header should give the same IP, but apparently not always (observed miniSipServer modifying Contact header, weird)